Add ability to test full action locally with act
The docker images that act uses to test GitHub actions locally don't
include GitHub CLI (unlike the virtual environments that GitHub actions
actually run in), so it wasn't previously possible to test the action
locally.

Now it is, as we're testing for the presence of the GITHUB_ACTIONS
environment variable, that all GitHub actions set to true by default
(https://docs.github.com/en/actions/reference/environment-variables).

If that variable isn't present, we install GitHub CLI, so it's available
locally for running the rest of the action.