Output what is happening during script run
We are seeing increasing, but intermittent, failures when running this
action on GitHub (though the script still seems to be completing
successfully when run locally).

The only error we're getting is `gh: Not Found (HTTP 404)`, which
suggests the script is making an API call to an endpoint that doesn't
exist. To work out exactly when this is happening, this commit adds more
output from the script, so we can tell what it's doing when it fails.