Remove the whippet-app topic from archived repos
Before: the script would add the "whippet-app" topic to found repos that
didn't already have it (unless they were archived), but if a repo was
archived after the "whippet-app" topic was added, we'd have to remember
to manually remove the topic.

Now: if a repo is found that is archived and has the "whippet-app"
topic, the "whippet-app" topic is removed. This will stop GovPress Tools
reporting on it.